--- src/de/Fraport/Feuerwehr/Enumerations/FahrzeugArt.java
@@ -1,9 +1,0 @@
-package de.Fraport.Feuerwehr.Enumerations;
-
-/**
- * Enumeration, die die möglichen Fahrzeugarten repräsentiert.
- * UNDEFINED gibt einen Fehler oder Abbruch an.
- */
-public enum FahrzeugArt {
-    PKW, LKW, UNDEFINED
-}
