+++ src/de/Fraport/Feuerwehr/Fahrzeug.java
@@ -1,9 +1,11 @@
 package de.Fraport.Feuerwehr;
 
-import de.Fraport.Feuerwehr.Enumerations.FahrzeugArt;
+import de.Fraport.Feuerwehr.Enumerations.Dienstgrad;
 import de.Fraport.Feuerwehr.Enumerations.Fahrzeugkategorie;
 
 import java.util.ArrayList;
+import java.util.Random;
+
 /**
  * Meistens Klasse, die ein allgemeines Fahrzeug repräsentiert.
  * Die Klasse stellt die folgenden Attribute bereit:
@@ -78,7 +80,7 @@ public class Fahrzeug {
      * @return Die FahrzeugKategorie
      */
     public Fahrzeugkategorie getFahrzeugKategorie(){
-        return fahrzeugKategorie;
+        return (Fahrzeugkategorie) fahrzeugKategorie;
     }
 
     /**
@@ -102,4 +104,137 @@ public class Fahrzeug {
         return rueckfahrendeBesatzung;
     }
 
+    /**
+     * Enumeration, die die möglichen Fahrzeugarten repräsentiert.
+     * UNDEFINED gibt einen Fehler oder Abbruch an.
+     */
+    enum FahrzeugArt {
+        PKW, LKW, UNDEFINED
+    }
+
+}
+/**
+ * Klasse, die einen Einsatzleitwagen (ELW) repräsentiert und von der Klasse Fahrzeug erbt.
+ * die Klasse hat zusätzlich das Attribut dienstgrad, welches im Einsatz den höchsten Dienstgrad der auf dem Fahrzeug befindlichen Feuerwehrleute angibt
+ *
+ */
+class ELW/*Einsatzleitwagen*/ extends Fahrzeug {
+
+    private Dienstgrad dienstgrad;
+
+    /**
+     * Konstruktor für die ELW-Klasse.
+     *
+     * @param fahrzeugnummer Die Nummer des ELWs.
+     */
+    public ELW(int fahrzeugnummer) {
+        super(fahrzeugnummer, Fahrzeugkategorie.ELW, FahrzeugArt.PKW, 2);
+
+    }
+
+    /**
+     * Gibt den Dienstgrad des höchstrangigen Feuerwehrmanns auf dem ELW zurück
+     *
+     * @return Der Dienstgrad höchstrangigen Feuerwehrmanns
+     */
+    public Dienstgrad getDienstgrad() {
+        return dienstgrad;
+    }
+    /**
+     * Setzt den Dienstgrad des höchstrangigen auf dem ELW befindlichen Feuerwehrmanns.
+     *
+     * @param dienstgrad Der zu setzende Dienstgrad.
+     */
+    public void setDienstgrad(Dienstgrad dienstgrad) {
+        this.dienstgrad = dienstgrad;
+    }
+}
+/**
+ * Klasse, die ein Tanklöschfahrzeug repräsentiert und von der Klasse Fahrzeug erbt.
+ */
+class TLF/*Tankloeschfahrzeug*/ extends Fahrzeug{
+    /**
+     * Der maximale Tankinhalt des Tanklöschfahrzeugs.
+     */
+    private int tank;
+
+    /**
+     * Konstruktor für die TLF-Klasse.
+     * Der Tankinhalt wird aus einer Auswahl von zwei Tankgrößen zufällig gewählt, die Tankgrößen kommen von realen Fahrzeugen der Feuerwehr Frankfurt und stellen Liter dar
+     * @param fahrzeugnummer Die Fahrzeugnummer des Tanklöschfahrzeugs.
+     */
+    public TLF(int fahrzeugnummer) {
+        super(fahrzeugnummer, Fahrzeugkategorie.TLF, FahrzeugArt.LKW, 4);
+        int[] moeglicheTanks = {3000,11000};
+        Random rand = new Random();
+        this.tank = moeglicheTanks[rand.nextInt(moeglicheTanks.length)];
+    }
+
+    /**
+     * Gibt den Tankinhalt des Tanklöschfahrzeugs zurück.
+     *
+     * @return Der Tankinhalt.
+     */
+    public int getTank() {
+        return tank;
+    }
+
+}
+
+/**
+ * Klasse, die eine Drehleiter mit Korb (DLK) repräsentiert und von der Klasse Fahrzeug erbt.
+ * die Klasse hat zusätzlich das Attribut hoehe, welches die maximale Einsatzhöhe der DLK angibt
+ */
+class DLK/*Drehleiter (mit Korb)*/ extends Fahrzeug {
+    private int hoehe;
+
+    /**
+     * Konstruktor für die DLK-Klasse.
+     *
+     * @param fahrzeugnummer Die Nummer der Drehleiter.
+     * @param hoehe Die Höhe der Drehleiter mit Korb.
+     */
+    public DLK(int fahrzeugnummer, int hoehe) {
+        super(fahrzeugnummer, Fahrzeugkategorie.DLK, FahrzeugArt.LKW,2);
+        this.hoehe = hoehe;
+
+    }
+    /**
+     * getter für das Attribut hoehe.
+     *
+     * @return gibt die Höhe der Drehleiter mit Korb zurück.
+     */
+    public int getHoehe() {
+        return hoehe;
+    }
+}
+
+/**
+ * Klasse, die ein Mannschaftstransportfahrzeug (MTF) repräsentiert und von der Klasse Fahrzeug erbt.
+ * die Klasse hat zusätzlich das Attribut baujahr, welches das Baujahr des MTFs angibt
+ */
+class MTF/*Mannschaftstransportfahrzeug*/ extends Fahrzeug {
+    /**
+     * Das Baujahr des Mannschaftstransportfahrzeugs.
+     */
+    private int baujahr;
+
+    /**
+     * Konstruktor für die MTF-Klasse.
+     *
+     * @param fahrzeugnummer Die Nummer des Fahrzeugs.
+     * @param baujahr        Das Baujahr des Mannschaftstransportfahrzeugs.
+     */
+    public MTF(int fahrzeugnummer,int baujahr) {
+        super(fahrzeugnummer, Fahrzeugkategorie.MTF, FahrzeugArt.PKW, 14);
+        this.baujahr = baujahr;
+    }
+    /**
+     * Gibt das Baujahr des Mannschaftstransportfahrzeugs zurück.
+     *
+     * @return Das Baujahr des Mannschaftstransportfahrzeugs.
+     */
+    public int getBaujahr() {
+        return baujahr;
+    }
 }
