--- src/de/Fraport/Feuerwehr/LkwFahrer.java
@@ -1,16 +1,0 @@
-package de.Fraport.Feuerwehr;
-
-/**
- * Klasse, die einen Feuerwehrmann mit der Spezialisierung als LKW-Fahrer repräsentiert.
- */
-public class LkwFahrer extends Feuerwehrmann{
-
-    /**
-     * Konstruktor für die LkwFahrer-Klasse.
-     *
-     * @param personalnummer Die Personalnummer des LKW-Fahrers.
-     */
-    public LkwFahrer(int personalnummer) {
-        super(personalnummer);
-    }
-}
