--- src/de/Fraport/Feuerwehr/PkwFahrer.java
@@ -1,16 +1,0 @@
-package de.Fraport.Feuerwehr;
-
-/**
- * Klasse, die einen Feuerwehrmann repräsentiert, der nur als PKW-Fahrer tätig ist.
- */
-public class PkwFahrer extends Feuerwehrmann{
-
-    /**
-     * Konstruktor für die PkwFahrer-Klasse.
-     *
-     * @param personalnummer Die Personalnummer des Feuerwehrmanns.
-     */
-    public PkwFahrer(int personalnummer) {
-        super(personalnummer);
-    }
-}
